Add DecompUtils
