+++ mlir/lib/Quantum/Transforms/GraphDecomposition/graph_decomposition.cpp
@@ -60,6 +60,7 @@
 #include "DGBuilder.hpp"
 #include "DGSolver.hpp"
 #include "DGTypes.hpp"
+#include "DecompUtils.hpp"
 
 #define DEBUG_TYPE "graph-decomposition"
 
@@ -233,7 +234,7 @@ struct GraphDecompositionPass : public impl::GraphDecompositionPassBase<GraphDec
         llvm::StringRef ruleName = rule.getName();
 
         // 1. Mandatory Attribute Check (Target Gate and Resources)
-        auto targetGateAttr = rule->getAttrOfType<StringAttr>("target_gate");
+        auto targetGateAttr = rule->getAttrOfType<StringAttr>(DecompUtils::target_gate_attr_name);
         auto resourcesAttr = rule->getAttrOfType<DictionaryAttr>("resources");
         if (!targetGateAttr) {
             llvm::errs() << "Cannot parse decomposition rule " << ruleName
@@ -328,7 +329,7 @@ struct GraphDecompositionPass : public impl::GraphDecompositionPassBase<GraphDec
         }
 
         WalkResult walkResult = module.walk([&](mlir::func::FuncOp func) {
-            if (func->hasAttr("target_gate")) {
+            if (func->hasAttr(DecompUtils::target_gate_attr_name)) {
                 if (userRuleNames.contains(func.getName())) {
                     if (failed(addRuleNode(func, ruleNodes))) {
                         return WalkResult::interrupt();
@@ -361,7 +362,7 @@ struct GraphDecompositionPass : public impl::GraphDecompositionPassBase<GraphDec
         // Add words from existing paulirot rules
         module.walk([&](mlir::func::FuncOp func) {
             // TODO: generalize this
-            if (func->hasAttr("target_gate")) {
+            if (func->hasAttr(DecompUtils::target_gate_attr_name)) {
                 if (func.getName().starts_with("paulirot_decomp_rule_")) {
                     addedWords.insert(func.getName().drop_front(21));
                 }
@@ -416,7 +417,7 @@ struct GraphDecompositionPass : public impl::GraphDecompositionPassBase<GraphDec
 
             mlir::func::FuncOp funcOp = outOp.get();
             outOp->setName((outOp->getName() + "_" + pauliWord).str()); // unique name per pauliword
-            funcOp->setAttr("target_gate", mlir::StringAttr::get(context, "paulirot" + pauliWord));
+            funcOp->setAttr(DecompUtils::target_gate_attr_name, mlir::StringAttr::get(context, "paulirot" + pauliWord));
 
             if (failed(addRuleNode(funcOp, ruleNodes))) {
                 return failure();
@@ -427,23 +428,10 @@ struct GraphDecompositionPass : public impl::GraphDecompositionPassBase<GraphDec
         return success();
     }
 
-    bool isInDecompRule(Operation *op)
-    {
-        while (auto parentOp = op->getParentOp()) {
-            if (auto funcOp = dyn_cast<func::FuncOp>(parentOp)) {
-                if (funcOp->hasAttr("target_gate")) {
-                    return true;
-                }
-            }
-            op = parentOp;
-        }
-        return false;
-    }
-
     void getOperators(std::vector<OperatorNode> &operators)
     {
         getOperation().walk([&](quantum::QuantumGate op) {
-            if (isInDecompRule(op)) {
+            if (DecompUtils::isInDecompRule(op)) {
                 return;
             }
             OperatorNode node;
